+++ src/windows/shellnode.cpp
@@ -1,29 +1,12 @@
 #include "shellnode.h"
 #include <qewindows/shell.h>
+#include <qewindows/shellnodedata.h>
+#include <qewindows/shellcache.h>
 
 namespace qe {
 namespace windows {
 
-//! \brief Constructs a new instance from a `ShellItemPointer` and its parent node.
-//! Node that `parent` may be nullptr, indicating this is the desktop (root) node.
-ShellNode::ShellNode(ShellItemPointer item, ShellNodePointer parent)
-{
-    d.parent = parent;
-    d.data = ShellNodeData::create(item);
-}
-
-ShellNode::ShellNode(const IdListPointer &id, ShellNodePointer parent)
-{
-    d.parent = parent;
-    d.data = ShellNodeData::create(id);
-}
-
-ShellNodePointer ShellNode::parent() const noexcept
-{
-    return d.parent ? d.parent : nullptr;
-}
-
-bool ShellNode::hasChildren() const
+bool ShellNode::hasChildren() const noexcept
 {
     if (!isValid())
         return false;
@@ -52,6 +35,29 @@ void ShellNode::enumerate()
     }
     auto sf = d.data->item.queryInterface<IShellFolder>();
 
+}
+
+//! \brief Constructs a new instance from a `ShellItemPointer` and its parent node.
+//! Node that `parent` may be nullptr, indicating this is the desktop (root) node.
+ShellNode::ShellNode(ShellItemPointer item, ShellNodePointer parent, ShellCache *cache)
+{
+    Q_ASSERT(item);
+    if (!cache)
+        d.cache = cache;
+    else
+        d.cache = ShellCache::globalInstance();
+    d.key = ShellCache::keyFor(item.get());
+    Q_ASSERT(!d.key.isNull());
+    Q_ASSERT(!cache->contains(d.key));
+    d.data = ShellNodeData::create(item);
+    d.parent = parent;
+    d.cache->insert(ShellNodePointer(this));
+}
+
+ShellNode::ShellNode(const IdListPointer &id, ShellNodePointer parent, ShellCache *cache)
+{
+    d.parent = parent;
+    d.data = ShellNodeData::create(id);
 }
 
 } // namespace windows