+++ test/windows/main.cpp
@@ -1,5 +1,6 @@
 #include "test_unknownpointer.h"
 #include "test_shellnodeinfo.h"
+#include "test_shellcache.h"
 
 int main(int argc, char *argv[])
 {
@@ -10,6 +11,8 @@ int main(int argc, char *argv[])
 
     unknown_pointer_test::run();
     shell_node_info_test::run();
+    shell_cache_test::run();
+
     CoUninitialize();
     return 0;
 }