+++ changelog.md
@@ -1,8 +1,16 @@
 # Changelog
 
+### 2018-06-07
+* qewindows/shell: added `createBindContext()`
+* Created test_shellcache.h in the test project.
+* Lots of fixes.
+* Still working at auto-populating the cache on ShellNode creation.
+
+### 2018-06-05
+* qewindows/shellcache: worked on storing data as a `QMap` instead of a `QHash`.
+
 ### 2018-06-04
 * qewindows/shellcache: added `keyFor` for parsing pathnames.
-* qewindows/shellnodedata: added `clear` function. This may be at the wrong level.
 
 ### 2018-06-03
 * qecore/type_util: cleaned up and removed everything from the ::detail namespace.